feat: [ANDROAPP-5125] Refactor profile picture path extension to Tracker module
