+++ app/src/main/java/org/dhis2/usescases/enrollment/EnrollmentModule.kt
@@ -39,7 +39,7 @@ import org.dhis2.form.ui.provider.KeyboardActionProviderImpl
 import org.dhis2.form.ui.provider.LegendValueProviderImpl
 import org.dhis2.form.ui.provider.UiEventTypesProviderImpl
 import org.dhis2.form.ui.validation.FieldErrorMessageProvider
-import org.dhis2.usescases.teiDashboard.TeiAttributesProvider
+import org.dhis2.tracker.relationships.data.TeiAttributesProvider
 import org.dhis2.utils.analytics.AnalyticsHelper
 import org.hisp.dhis.android.core.D2
 import org.hisp.dhis.android.core.arch.repositories.`object`.ReadOnlyOneObjectRepositoryFinalImpl
