+++ app/src/main/java/org/dhis2/usescases/enrollment/EnrollmentPresenterImpl.kt
@@ -3,7 +3,6 @@ package org.dhis2.usescases.enrollment
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.processors.FlowableProcessor
 import io.reactivex.processors.PublishProcessor
-import org.dhis2.bindings.profilePicturePath
 import org.dhis2.commons.bindings.trackedEntityTypeForTei
 import org.dhis2.commons.data.TeiAttributesInfo
 import org.dhis2.commons.matomo.Actions.Companion.CREATE_TEI
@@ -14,7 +13,8 @@ import org.dhis2.commons.schedulers.SchedulerProvider
 import org.dhis2.commons.schedulers.defaultSubscribe
 import org.dhis2.form.data.EnrollmentRepository
 import org.dhis2.form.model.RowAction
-import org.dhis2.usescases.teiDashboard.TeiAttributesProvider
+import org.dhis2.tracker.extensions.profilePicturePath
+import org.dhis2.tracker.relationships.data.TeiAttributesProvider
 import org.dhis2.utils.analytics.AnalyticsHelper
 import org.dhis2.utils.analytics.DELETE_AND_BACK
 import org.hisp.dhis.android.core.D2
